clean deployed files
We typically make PR changes that add and delete files all the time.
If we don't clean up the repo we store them in then templates that
we delete thru PR will never get removed from the repo.  This
fixes that situation by removing existing templates and replacing
them with the latest files.