add a bucket URL (#62)
SC end users don’t have list bucket permission so they can’t just goto the S3
console and type the bucket name to access their bucket.  However they can acess
the bucket with a direct link.  We provide a direct link in the outputs
to make it easier to navigate to the bucket from the provisioned
products list page.