Remove taskcat (#112)
Just came upon a situation where we added cfn-sc-actions-provider
custom resources to templates in scipoolprod-sc-lib.  The tests
actually run ogainst admincentral so they fail because the
underlying resources aren't available in admincental.  The only
way to make them work is to add a bunch of infrastructure to
admincentral only for the purpose of running taskcat tests.
That just didn't seem like it's worth it to me.

Also taskcat hasn't been all that valuable to us since we don't
maintain the tests and don't write tests for new templates
we add to this repo.

I've officially given up on taskcat.