[SC-241] Fix EC2 tagging policy
The set_env_vars script and related policies are used to facilitate tagging
resources from an EC2 instance.  Now that we have moved to using the
cfn-cr-synapse-tagger[1] to do all the tagging work we can remove that script
and all it's associated policies.  This will make the template simpler
and make provisioning instances faster.

[1] https://github.com/Sage-Bionetworks-IT/cfn-cr-synapse-tagger