make sure bucket policy depends on bucket
A bucket delete may fail due to having existing data in the bucket.
When a failure occurs the bucket will remain however the  policy for it
may get removed.  This will cause a state where users will no longer
have access to the bucket.

This change makes sure that the policy deletion only happens after a
bucket has been delete.