[SC-326] Retain bucket on SC product termination (#259)
A bucket delete may fail due to having existing data in the bucket.
When a failure occurs the bucket will remain and SC will report that
the product is in a `Tainted` state.  When this occurs users cannot update
or terminate the product again which means they are stuck requiring
an AWS admin to help resolve the situation.

This change will make cloudformation ignore deleting the bucket.
When a user terminates the SC product the policy on the bucket
and the SC product is removed however the bucket itself is retained.
CFN doesn't even try to remove the bucket so it won't ever get into
the `tainted` state.  The bucket stays around and the policy is
removed so nobody has access to the bucket after an SC termination.