[SC-306] Fix Command parameter
The scheuled jobs product failed to deploy because the ContainerProperties
Command expects a list of strings.

```
Property validation failure: [Value of property {/ContainerProperties/Command} does not match type {Array}]
```

To fix that we need to use the CFN split function to create a list from
a space delimited string of commands.