[IT-3222] Only allow AMD instances
We have decided that the best option to balance performance, compability
and cost would be to only allow the use of AMD instances in the
service catalog.  This change restricts the list of instances users
get to select when provisioning SC EC instances.  It also attempts to
reduce the number of instance types that we support by removing
redundant types.