[IT-3222] Only allow AMD instances (#311)
We have decided that the best option to balance performance, compability and cost would be to only allow the use of AMD instances[1] in the service catalog.  This change restricts the list of instances users get to select when provisioning service catalog EC2 instances.  It also attempts to reduce the number of instance types that we support by removing redundant types.

[1] https://aws.amazon.com/ec2/instance-types/

depends on #312